+++ server/pipeline_match.go
@@ -207,16 +207,12 @@ func (p *Pipeline) matchJoin(logger *zap.Logger, session Session, envelope *rtap
 		}
 		if mode == StreamModeMatchAuthoritative {
 			// If we've reached here, it was an accepted authoritative join.
-			ctx, ctxCancelFn := context.WithTimeout(session.Context(), 5*time.Second)
+			ctx, ctxCancelFn := context.WithTimeout(session.Context(), time.Duration(p.config.GetMatch().JoinMarkerDeadlineMs)*time.Millisecond)
 			if err := p.matchRegistry.AwaitJoinMarker(ctx, matchID, node, session.ID(), p.node); err != nil {
-				if err != context.Canceled {
-					ctxCancelFn()
-				}
 				// There was an error or a timeout waiting for the join marker, return to the client anyway since the tracker update was successful.
 				logger.Error("Error waiting for match join marker", zap.Error(err))
-			} else {
-				ctxCancelFn()
 			}
+			ctxCancelFn()
 			label = &wrappers.StringValue{Value: l}
 		}
 		meta = &m